fix: Backend structure code refactor
