--- backend/tools/__init__.py
@@ -1,3 +1,0 @@
-from WRS2Tools import WRS2Tools
-
-WRS2Tools = WRS2Tools
