+++ backend/utils/__init__.py
@@ -1,0 +1,3 @@
+from GeoLocationUtils import WRS2Utils
+
+WRS2Tools = WRS2Utils
